+++ qrenderdoc/Windows/PipelineState/D3D12PipelineStateViewer.cpp
@@ -1733,7 +1733,7 @@ void D3D12PipelineStateViewer::resource_itemActivated(RDTreeWidgetItem *item, in
     if(view.res.Resource != ResourceId())
     {
       offs = view.res.FirstElement * view.res.ElementSize;
-      size = view.res.NumElements * view.res.ElementSize;
+      size = uint64_t(view.res.NumElements) * view.res.ElementSize;
     }
     else
     {
